pyproject.toml
[tool.poetry]
name = "definitionapi"
version = "0.1.0"
description = ""
authors = ["Isai <isairamirez15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
